Split oversized runtime contract tests
