+++ packages/runtime/src/contracts/root-api.test.ts
@@ -15,26 +15,14 @@ import type {
 } from "../execution";
 import type {
   AgentAutoCompactionOptions,
-  AgentEvent,
   AgentHost,
   AgentInput,
   AgentOptions,
-  AgentPluginInterceptResult,
-  BeforeToolCall,
-  ControlAgentEvent,
-  LifecycleAgentEvent,
-  TelemetryAgentEvent,
   ThreadCompactionInput,
   ThreadHandle,
-  VisibleAgentEvent,
 } from "../index";
 import {
   Agent,
-  isBeforeToolCallEvent,
-  isControlAgentEvent,
-  isLifecycleAgentEvent,
-  isTelemetryAgentEvent,
-  isVisibleAgentEvent,
   runPluginsForEvent,
   threadStoreKey as runtimeThreadStoreKey,
 } from "../index";
@@ -71,10 +59,6 @@ describe("runtime public exports", () => {
 
     expect(runtime).toHaveProperty("Agent", Agent);
     expect(runtime).toHaveProperty("runPluginsForEvent", runPluginsForEvent);
-    expect(runtime).toHaveProperty(
-      "isBeforeToolCallEvent",
-      isBeforeToolCallEvent
-    );
     expect(runtime).not.toHaveProperty("runPluginsForToolCall");
     expect(runtime).toHaveProperty("threadStoreKey", runtimeThreadStoreKey);
     expect(runtime).not.toHaveProperty("createInMemoryExecutionHost");
@@ -92,21 +76,6 @@ describe("runtime public exports", () => {
     expect(emptyHostIsRejected).toBe(false);
   });
 
-  it("exports event classifiers from the package root", async () => {
-    const runtime = await import("../index");
-
-    expect(runtime).toHaveProperty("isVisibleAgentEvent", isVisibleAgentEvent);
-    expect(runtime).toHaveProperty(
-      "isLifecycleAgentEvent",
-      isLifecycleAgentEvent
-    );
-    expect(runtime).toHaveProperty(
-      "isTelemetryAgentEvent",
-      isTelemetryAgentEvent
-    );
-    expect(runtime).toHaveProperty("isControlAgentEvent", isControlAgentEvent);
-  });
-
   it("does not expose runtime-owned subagent helpers from the package root", async () => {
     const runtime = await import("../index");
 
@@ -121,32 +90,6 @@ describe("runtime public exports", () => {
     for (const forbiddenName of forbiddenModelAdapterRootExports) {
       expect(runtime).not.toHaveProperty(forbiddenName);
     }
-  });
-
-  it("types event classifier exports from the package root", () => {
-    const visible = {
-      text: "hello",
-      type: "assistant-output",
-    } satisfies VisibleAgentEvent;
-    const lifecycle = { type: "turn-start" } satisfies LifecycleAgentEvent;
-    const telemetry = {
-      text: "thinking",
-      type: "assistant-reasoning",
-    } satisfies TelemetryAgentEvent;
-    const control = lifecycle satisfies ControlAgentEvent;
-    const events = [visible, lifecycle, telemetry, control] satisfies readonly [
-      VisibleAgentEvent,
-      LifecycleAgentEvent,
-      TelemetryAgentEvent,
-      Exclude<AgentEvent, VisibleAgentEvent>,
-    ];
-
-    expect(events.map((event) => event.type)).toEqual([
-      "assistant-output",
-      "turn-start",
-      "assistant-reasoning",
-      "turn-start",
-    ]);
   });
 
   it("types public thread compaction options from the package root", () => {
@@ -187,29 +130,6 @@ describe("runtime public exports", () => {
     expect(enabledOptions.notificationOverlays).toEqual(["runtime context"]);
     expect(disabledOptions.autoCompaction).toBe(false);
     expect(compaction.startSeq).toBe(0);
-  });
-
-  it("types before-tool-call plugin interception results", () => {
-    const event = {
-      attempt: 1,
-      idempotencyKey: "run-1:call_tool-1",
-      input: { path: "README.md" },
-      policy: "manual-recovery",
-      toolCallId: "call_tool-1",
-      toolName: "write_file",
-      type: "before-tool-call",
-    } satisfies BeforeToolCall;
-    const continueResult = {
-      action: "continue",
-    } satisfies AgentPluginInterceptResult;
-    const recoveryResult = {
-      action: "needs-recovery",
-    } satisfies AgentPluginInterceptResult;
-
-    expect(event.toolName).toBe("write_file");
-    expect(isBeforeToolCallEvent(event)).toBe(true);
-    expect(continueResult.action).toBe("continue");
-    expect(recoveryResult.action).toBe("needs-recovery");
   });
 
   it("types advanced host contracts", () => {
